backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
requests>=2.31.0
pydantic>=2.9.0
sentence-transformers>=2.7.0
faiss-cpu>=1.12.0
numpy>=1.26.4
python-dotenv>=1.0.1
